feat: collapse catalog.js + catalog-typed.ts into single catalog.ts
Single catalog module owning state, registration, promises, and the typed
Result-returning consumer API. Deletes the legacy isXReady predicates and
unsafe getters that lingered through commits 1-3, then merges the two
files into one TypeScript module so imports go to one place.

Source-side:
  - sources/state/catalog.ts (new): contains everything — types, stages,
    stores (private), catalogReady promises, register*, test utilities,
    Result-returning getters, chunkReady boundary helper, isXReady
    predicates, globalThis shims for Playwright/Argos.
  - sources/state/catalog.js (deleted), catalog-typed.ts (deleted).

Public API surface:
  - catalogReady (async chunk-ready promises)
  - isIndexReady() / isLiteReady() / isCreditsReady() / isPaletteReady()
    / isLayersReady() — sync boolean predicates (encapsulated; expose
    only the readiness, not the underlying Stage object)
  - chunkReady(chunk) — Result<true, LoadError> primitive for
    Result.combine boundary patterns
  - getX(id) / getX() — typed Result-returning getters (8)
  - registerFromXModule (5) — chunk loaders called by install-item-metadata
  - buildItemsByTypeNameFromRegisteredLite — hash.js helper
  - loadCatalogFromFixtures, resetCatalogForTests — test utilities
  - All consumer-facing types: ItemLite, ItemMerged, Credit, PaletteRecolor,
    PaletteMetadata, AliasMetadata, MetadataIndexes, etc.

What's intentionally NOT exported anymore:
  - stages object: was leaking the underlying Stage (promise + resolver)
    via .resolved field. Replaced by isXReady() functions that return
    just the boolean.
  - stores object: was leaking module state for test inspection. Tests
    now exercise the public API (Result-returning getters).

Architectural debt that remains (TODO above resetCatalogForTests):
  module-level singleton state. The real fix is a createCatalog() factory
  + DI; that would eliminate resetCatalogForTests entirely. Deferred —
  out of scope for this migration.

Test changes:
  - tests/state/catalog_spec.js: rewritten to use the public API. Asserts
    isXReady() flips and Result variants instead of poking stores/stages
    directly.
  - tests/state/catalog-typed_spec.js renamed -> catalog-getters_spec.js
    (the file is no longer "the typed wrapper spec" since there's no
    separate wrapper).
  - tests/components/tree/PaletteSelectModal_spec.js: stages.X.resolved
    -> isXReady() in assertions.
  - tests/components/tree/{ItemWithRecolors,ItemWithVariants}_spec.js,
    tests/state/palettes_spec.js: catalog.getItemMerged/getItemLite ->
    typed-getter equivalents.
  - All `from "./catalog.js"` and `from "./catalog-typed.ts"` import
    paths updated to `./catalog.ts` (Vite's bundler resolution would
    have handled .js -> .ts, but Node's strict ESM in test:node would
    not, so consistency wins).

Verified: type-check, lint, 118 node tests, 431 browser tests, build.
