+++ sources/components/download/Download.js
@@ -17,9 +17,9 @@ import {
   exportIndividualFrames,
 } from "../../state/zip.js";
 import { debugLog } from "../../utils/debug.js";
-import { stages } from "../../state/catalog.js";
+import { isLayersReady } from "../../state/catalog.ts";
 
-const zipExportDisabled = () => !stages.layers.resolved;
+const zipExportDisabled = () => !isLayersReady();
 const zipExportTitle = "Wait for layer data to finish loading";
 
 export const Download = {
