+++ sources/state/state.js
@@ -2,7 +2,7 @@
 import m from "mithril";
 import { LICENSE_CONFIG, ANIMATIONS, BODY_TYPES } from "./constants.ts";
 import { syncSelectionsToHash, loadSelectionsFromHash } from "./hash.js";
-import { getItemMerged } from "./catalog-typed.ts";
+import { getItemMerged } from "./catalog.ts";
 import { renderCharacter } from "../canvas/renderer.js";
 
 // Dependency injection for testability (see setStateDeps / resetStateDeps)
