+++ tests/components/selections/CurrentSelections_spec.js
@@ -3,7 +3,7 @@ import { assert } from "chai";
 import { describe, it, beforeEach, afterEach } from "mocha-globals";
 import { CurrentSelections } from "../../../sources/components/selections/CurrentSelections.js";
 import { state } from "../../../sources/state/state.js";
-import { resetCatalogForTests } from "../../../sources/state/catalog.js";
+import { resetCatalogForTests } from "../../../sources/state/catalog.ts";
 import {
   resetState,
   setEnabledLicenses,
