+++ tests/state/hash_spec.js
@@ -19,7 +19,7 @@ import {
   resetHashCalledTimes,
   resetHashDeps,
 } from "../../sources/state/hash.js";
-import { resetCatalogForTests } from "../../sources/state/catalog.js";
+import { resetCatalogForTests } from "../../sources/state/catalog.ts";
 import {
   restoreAppCatalogAfterTest,
   seedBrowserCatalog,
