+++ tests/tests.js
@@ -30,7 +30,7 @@ import "./components/preview/PreviewMetadataLoadingOverlay_spec.js";
 import "./components/preview/PinchToZoom_spec.js";
 import "./components/selections/CurrentSelections_spec.js";
 import "./state/catalog_spec.js";
-import "./state/catalog-typed_spec.js";
+import "./state/catalog-getters_spec.js";
 import "./state/filters_spec.js";
 import "./state/hash_spec.js";
 import "./state/json_spec.js";
